Move CheckoutInfo and CheckoutLine info logic and dataclasses to fetch file (#6907)
* Move CheckoutInfo and CheckoutLine info logic and dataclasses to fetch file in checkout

* Fix lgtm import alert
