+++ saleor/checkout/__init__.py
@@ -1,18 +1,4 @@
 import logging
-from dataclasses import dataclass
-from typing import TYPE_CHECKING, List, Optional
-
-if TYPE_CHECKING:
-    from ..account.models import Address, User
-    from ..channel.models import Channel
-    from ..product.models import (
-        Collection,
-        Product,
-        ProductVariant,
-        ProductVariantChannelListing,
-    )
-    from ..shipping.models import ShippingMethod, ShippingMethodChannelListing
-    from .models import Checkout, CheckoutLine
 
 logger = logging.getLogger(__name__)
 
@@ -25,30 +11,3 @@ class AddressType:
         (BILLING, "Billing"),
         (SHIPPING, "Shipping"),
     ]
-
-
-@dataclass
-class CheckoutLineInfo:
-    line: "CheckoutLine"
-    variant: "ProductVariant"
-    channel_listing: "ProductVariantChannelListing"
-    product: "Product"
-    collections: List["Collection"]
-
-
-@dataclass
-class CheckoutInfo:
-    checkout: "Checkout"
-    user: Optional["User"]
-    channel: "Channel"
-    billing_address: Optional["Address"]
-    shipping_address: Optional["Address"]
-    shipping_method: Optional["ShippingMethod"]
-    valid_shipping_methods: List["ShippingMethod"]
-    shipping_method_channel_listings: Optional["ShippingMethodChannelListing"]
-
-    def get_country(self) -> str:
-        address = self.shipping_address or self.billing_address
-        if address is None or not address.country:
-            return self.checkout.country.code
-        return address.country.code
