+++ saleor/graphql/checkout/tests/test_checkout.py
@@ -14,18 +14,15 @@ from prices import Money, TaxedMoney
 
 from ....account.models import User
 from ....channel.utils import DEPRECATION_WARNING_MESSAGE
-from ....checkout import CheckoutInfo, calculations
+from ....checkout import calculations
 from ....checkout.checkout_cleaner import (
     clean_checkout_payment,
     clean_checkout_shipping,
 )
 from ....checkout.error_codes import CheckoutErrorCode
+from ....checkout.fetch import CheckoutInfo, fetch_checkout_info, fetch_checkout_lines
 from ....checkout.models import Checkout
-from ....checkout.utils import (
-    add_variant_to_checkout,
-    fetch_checkout_info,
-    fetch_checkout_lines,
-)
+from ....checkout.utils import add_variant_to_checkout
 from ....core.payments import PaymentInterface
 from ....payment import TransactionKind
 from ....payment.interface import GatewayResponse
